Add py.typed file so that typecheckers recognize this library as typed
This is necessary as type-checking tools like mypy, pyright and ty use
this file as described by PEP-0561 to recognize typed libraries. This
is useful for better DX where, when types are available, they may
enhance autocompletion and/or provide early warnings of invalid code.

I would also recommend adding, if possible, types to the non-generated
parts of the code (like the GhApi class for example).

Closes #178.